Declare JKRHeap operator new with size_t

+++ libs/JSystem/src/JKernel/JKRHeap.cpp
@@ -252,28 +252,28 @@ void JKRHeap::fillMemory(u8* dst, u32 size, u8 value)
 	JUT_ASSERT_F(false, "UNIMPLEMENTED");
 }
 
-void* operator new(u32 byteCount)
+void* operator new(size_t byteCount)
 {
 	return JKRHeap::alloc(byteCount, 4, nullptr);
 }
-void* operator new(u32 byteCount, int alignment)
+void* operator new(size_t byteCount, int alignment)
 {
 	return JKRHeap::alloc(byteCount, alignment, nullptr);
 }
-void* operator new(u32 byteCount, JKRHeap* heap, int alignment)
+void* operator new(size_t byteCount, JKRHeap* heap, int alignment)
 {
 	return JKRHeap::alloc(byteCount, alignment, heap);
 }
 
-void* operator new[](u32 byteCount)
+void* operator new[](size_t byteCount)
 {
 	return JKRHeap::alloc(byteCount, 4, nullptr);
 }
-void* operator new[](u32 byteCount, int alignment)
+void* operator new[](size_t byteCount, int alignment)
 {
 	return JKRHeap::alloc(byteCount, alignment, nullptr);
 }
-void* operator new[](u32 byteCount, JKRHeap* heap, int alignment)
+void* operator new[](size_t byteCount, JKRHeap* heap, int alignment)
 {
 	return JKRHeap::alloc(byteCount, alignment, heap);
 }
